Refactor ddd interface package + introducte readmodel / readmodelrepository
