+++ ddd/interface/__init__.py
@@ -1,0 +1,11 @@
+from .domain_models import *
+from .queries import *
+from .builders import *
+from .exceptions import *
+
+from decimal import Decimal
+from typing import Union, List, Callable
+
+ApplicationServiceResult = Union[EntityIdentity, List[EntityIdentity], DTO, List[DTO]]
+ApplicationService = Callable[[CommandRequest], ApplicationServiceResult]
+PrimitiveType = Union[int, str, float, Decimal]
